requirements.txt
fastapi
uvicorn
crewai
langchain
langchain-groq
pypdf
python-dotenv
streamlit
requests